# Add SSE2 version for webp Filter methods