… tests/ImageSharp.Tests/Formats/WebP/WebpDecoderTests.cs