### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description

This PR add SSE2 version of the filter methods which are used during lossy decoding. Those make a big part of the decoding time.

Related #1786

The following methods now have SSE2 versions:
- DoFilter2, DoFilter4, DoFilter6
- HFilter16, VFilter16
- VFilter16i, HFilter16i
- VFilter8, HFilter8
- VFilter8i, HFilter8i
- SimpleHFilter16, SimpleVFilter16
- SimpleVFilter16i, SimpleHFilter16i

Benchmark results:

### Master:
```
|                     Method |        TestImageLossy |        TestImageLossless |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |---------------------- |------------------------- |---------:|---------:|---------:|------:|------:|------:|----------:|
|        'Magick Lossy Webp' | Webp/earth_lossy.webp | Webp/earth_lossless.webp | 114.3 ms | 11.16 ms |  6.64 ms |     - |     - |     - |     16 KB |
|    'ImageSharp Lossy Webp' | Webp/earth_lossy.webp | Webp/earth_lossless.webp | 264.8 ms | 13.15 ms |  7.83 ms |     - |     - |     - |  1,755 KB |
|     'Magick Lossless Webp' | Webp/earth_lossy.webp | Webp/earth_lossless.webp | 109.4 ms |  4.31 ms |  2.57 ms |     - |     - |     - |     16 KB |
| 'ImageSharp Lossless Webp' | Webp/earth_lossy.webp | Webp/earth_lossless.webp | 249.0 ms | 17.71 ms | 10.54 ms |     - |     - |     - |  1,735 KB |
```

### PR
```
|                     Method |        TestImageLossy |        TestImageLossless |     Mean |    Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |---------------------- |------------------------- |---------:|---------:|--------:|------:|------:|------:|----------:|
|        'Magick Lossy Webp' | Webp/earth_lossy.webp | Webp/earth_lossless.webp | 106.4 ms |  3.89 ms | 2.31 ms |     - |     - |     - |     16 KB |
|    'ImageSharp Lossy Webp' | Webp/earth_lossy.webp | Webp/earth_lossless.webp | 181.3 ms |  1.86 ms | 1.10 ms |     - |     - |     - |  1,755 KB |
|     'Magick Lossless Webp' | Webp/earth_lossy.webp | Webp/earth_lossless.webp | 107.5 ms |  3.95 ms | 2.35 ms |     - |     - |     - |     16 KB |
| 'ImageSharp Lossless Webp' | Webp/earth_lossy.webp | Webp/earth_lossless.webp | 237.2 ms | 10.20 ms | 6.07 ms |     - |     - |     - |  1,736 KB |
```